Add files via upload
This program is able to enter the customers or vehicles’ 
information and save it .When the user order the data, the 
program can print all of the content in the screen, or exit the 
program. The user can input a different type such as integer, 
character, and string data. It has a menu so the user can 
choose from one to four.